AddTokenPage refactor (#129)
* feat: created token folder, move all token files to that folder. Refactored add token page

* feat: refactoring completed. Added logic to avoid editing decimals when adding custom token

* feat: Added useSubmitOnEnter for both views and for confirm page

* feat: fix margin top for customView

* feat: Rename components. Refactor AddTokenManualView to remove useResult and keep useForm

* feat: fix merge and validate if decimals !=0 cant modify it

* feat: fix component name in ManualView, fix validations, remove some duplicated. Remove onChangeSymbol and add useEffect to check changes

* feat: Add useAccountToken hook. Remove unnecesary try-catch. Matching component names with file name. Fix message typo. Fix validation logic for manual view

* feat: fix message and merging validation in one IF clause. Changin onChange to sync and removing .then to await for the result of the searchedvalue

* feat: Remove .then when checking manual token data

* feat: Remove unnecessary else clause

* feat: change res name to tokenSearchResponse and remove if
